+++ frontend/components/decorators/DecoratorsPageWrapper/DecoratorsPageWrapper.jsx
@@ -1,0 +1,19 @@
+import { Component, PropTypes } from 'react';
+
+class DecoratorsPageWrapper extends Component {
+  static propTypes = {
+    children: PropTypes.node,
+  };
+
+  render() {
+    const { children } = this.props;
+
+    if (!children) {
+      return false;
+    }
+
+    return children;
+  }
+}
+
+export default DecoratorsPageWrapper;