+++ frontend/kolide/index.js
@@ -15,6 +15,7 @@ import statusLabelMethods from 'kolide/entities/status_labels';
 import targetMethods from 'kolide/entities/targets';
 import userMethods from 'kolide/entities/users';
 import websocketMethods from 'kolide/websockets';
+import decoratorMethods from 'kolide/entities/decorators';
 
 const DEFAULT_BODY = JSON.stringify({});
 
@@ -37,6 +38,7 @@ class Kolide extends Base {
     this.targets = targetMethods(this);
     this.users = userMethods(this);
     this.websockets = websocketMethods(this);
+    this.decorators = decoratorMethods(this);
   }
 
   authenticatedDelete (endpoint, overrideHeaders = {}) {