+++ frontend/redux/nodes/entities/reducer.js
@@ -9,6 +9,7 @@ import packs from './packs/reducer';
 import queries from './queries/reducer';
 import scheduledQueries from './scheduled_queries/reducer';
 import users from './users/reducer';
+import decorators from './decorators/reducer';
 
 export default combineReducers({
   campaigns,
@@ -20,4 +21,5 @@ export default combineReducers({
   queries,
   scheduled_queries: scheduledQueries,
   users,
+  decorators,
 });