+++ frontend/router/paths.js
@@ -14,6 +14,7 @@ export default {
   MANAGE_HOSTS: '/hosts/manage',
   NEW_PACK: '/packs/new',
   NEW_QUERY: '/queries/new',
+  NEW_DECORATOR: '/decorators/new',
   RESET_PASSWORD: '/login/reset',
   SETUP: '/setup',
   USER_SETTINGS: '/settings',